--- a/onnxruntime/core/providers/cpu/sequence/concat_from_sequence.cc
+++ b/onnxruntime/core/providers/cpu/sequence/concat_from_sequence.cc
@@ -40,7 +40,7 @@ Status ConcatFromSequence::Compute(OpKernelContext* ctx) const {
     return Status::OK();
 
   // Compute values to be placed in the output tensor
-  return ComputeImpl(p);
+  return ComputeImpl(p, ctx);
 }
 
 }  // namespace onnxruntime