--- a/onnxruntime/core/providers/cpu/tensor/copy.cc
+++ b/onnxruntime/core/providers/cpu/tensor/copy.cc
@@ -25,21 +25,46 @@ Status DispatchStridedCopy(concurrency::ThreadPool* thread_pool,
                            const TensorShape& copy_shape,
                            const Tensor& src,
                            const std::vector<int64_t> src_strides) {
-#define CALL_FOR_TYPE(T) \
-  StridedCopy<T>(thread_pool, dst.MutableData<T>() + dst_offset, copy_shape, dst_strides, src.Data<T>(), src_strides)
   ORT_RETURN_IF_NOT(dst.DataType() == src.DataType(), "src and dst types must match");
-  if (dst.IsDataType<float>()) {
-    CALL_FOR_TYPE(float);
-  } else if (dst.IsDataType<double>()) {
-    CALL_FOR_TYPE(double);
-  } else if (dst.IsDataType<int32_t>()) {
-    CALL_FOR_TYPE(int32_t);
-  } else if (dst.IsDataType<int64_t>()) {
-    CALL_FOR_TYPE(int64_t);
-  } else if (dst.IsDataTypeString()) {
-    CALL_FOR_TYPE(std::string);
-  } else {
-    return ORT_MAKE_STATUS(ONNXRUNTIME, FAIL, "unsupported data type");
+  ORT_RETURN_IF_NOT(dst_strides.size() == src_strides.size() && src_strides.size() == copy_shape.NumDimensions(), "src and dst must have same shape");
+
+  // Manual dispatching: DispatchOnTensorType doesn't work here because we need to pass the type to the MutableData call
+#define CALL_FOR_TYPE(T)                                                                                               \
+  StridedCopy<T>(thread_pool, dst.MutableData<T>() + dst_offset, dst_strides, copy_shape, src.Data<T>(), src_strides); \
+  break
+
+  auto tensor_type = dst.DataType()->AsPrimitiveDataType()->GetDataType();
+  switch (tensor_type) {
+    case ONNX_NAMESPACE::TensorProto_DataType_FLOAT:
+      CALL_FOR_TYPE(float);
+    case ONNX_NAMESPACE::TensorProto_DataType_BOOL:
+      CALL_FOR_TYPE(bool);
+    case ONNX_NAMESPACE::TensorProto_DataType_DOUBLE:
+      CALL_FOR_TYPE(double);
+    case ONNX_NAMESPACE::TensorProto_DataType_STRING:
+      CALL_FOR_TYPE(std::string);
+    case ONNX_NAMESPACE::TensorProto_DataType_INT8:
+      CALL_FOR_TYPE(int8_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_UINT8:
+      CALL_FOR_TYPE(uint8_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_INT16:
+      CALL_FOR_TYPE(int16_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_UINT16:
+      CALL_FOR_TYPE(uint16_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_INT32:
+      CALL_FOR_TYPE(int32_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_UINT32:
+      CALL_FOR_TYPE(uint32_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_INT64:
+      CALL_FOR_TYPE(int64_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_UINT64:
+      CALL_FOR_TYPE(uint64_t);
+    case ONNX_NAMESPACE::TensorProto_DataType_FLOAT16:
+      CALL_FOR_TYPE(MLFloat16);
+    case ONNX_NAMESPACE::TensorProto_DataType_BFLOAT16:
+      CALL_FOR_TYPE(BFloat16);
+    default:
+      ORT_ENFORCE(false, "Unknown tensor type of ", tensor_type);
   }
   return Status::OK();
 }
@@ -47,27 +72,34 @@ Status DispatchStridedCopy(concurrency::ThreadPool* thread_pool,
 template <typename T>
 void StridedCopy(concurrency::ThreadPool* thread_pool,
                  T* dst,
-                 const TensorShape& dst_shape,
                  const std::vector<int64_t>& dst_strides,
+                 const TensorShape& copy_shape,
                  const T* src,
                  const std::vector<int64_t>& src_strides) {
   const auto* src_raw = reinterpret_cast<const uint8_t*>(src);
   auto* dst_raw = reinterpret_cast<uint8_t*>(dst);
-  const size_t dims = dst_shape.NumDimensions();
+  const size_t dims = copy_shape.NumDimensions();
   // We will iterate over the output dimensions
   int64_t num_iterations = 1;
   for (size_t dim = 0; dim < dims; dim++) {
-    num_iterations *= dst_shape[dim];
+    num_iterations *= copy_shape[dim];
   }
 
-  // TODO(orausch): Reorder dimensions so that we iterate along the smallest strides first
-  // TODO(orausch): remove single size dimensions
+  if (num_iterations == 1) {
+    // scalar edge case
+    dst[0] = src[0];
+    return;
+  }
+
+  // TODOs for when we have strided tensors:
+  // - Reorder dimensions so that we iterate along the smallest strides first
+  // - remove single size dimensions
   concurrency::ThreadPool::TryParallelFor(
       thread_pool, num_iterations,
       {static_cast<float>(sizeof(T)), static_cast<float>(sizeof(T)), 1.0F},
-      [dst_shape, dst_strides, dst, dst_raw, src, src_raw, src_strides, dims](std::ptrdiff_t first, std::ptrdiff_t last) {
+      [copy_shape, dst_strides, dst, dst_raw, src, src_raw, src_strides, dims](std::ptrdiff_t first, std::ptrdiff_t last) {
         // Compute the initial n-dimensional index and addresses
-        auto last_dim_size = dst_shape[dims - 1];
+        auto last_dim_size = copy_shape[dims - 1];
         auto last_dst_stride = dst_strides[dims - 1];
         auto last_src_stride = src_strides[dims - 1];
         bool is_string = std::is_same<T, std::string>::value;
@@ -75,7 +107,7 @@ void StridedCopy(concurrency::ThreadPool* thread_pool,
         {
           int64_t current_index = first;
           for (size_t dim = dims; dim > 0; dim--) {
-            auto shape_val = dst_shape[dim - 1];
+            auto shape_val = copy_shape[dim - 1];
             // Iterate from dims to 1 so we don't roll over to positive on the bounds check
             current_nd_idx[dim - 1] = current_index % shape_val;
             current_index /= shape_val;
@@ -108,7 +140,7 @@ void StridedCopy(concurrency::ThreadPool* thread_pool,
 
           // update the current_nd_idx if needed
           size_t dim = dims - 1;
-          while (dim > 0 && current_nd_idx[dim] >= dst_shape[dim]) {
+          while (dim > 0 && current_nd_idx[dim] >= copy_shape[dim]) {
             current_nd_idx[dim] = 0;
             dim--;
             current_nd_idx[dim]++;
@@ -116,17 +148,4 @@ void StridedCopy(concurrency::ThreadPool* thread_pool,
         }
       });
 }
-
-#define STRIDED_COPY_IMPL(T)                                            \
-  template void StridedCopy<T>(concurrency::ThreadPool * thread_pool,   \
-                               T * dst,                                 \
-                               const TensorShape& dst_shape,            \
-                               const std::vector<int64_t>& dst_strides, \
-                               const T* src,                            \
-                               const std::vector<int64_t>& src_strides);
-
-STRIDED_COPY_IMPL(int32_t)
-STRIDED_COPY_IMPL(int64_t)
-STRIDED_COPY_IMPL(float)
-STRIDED_COPY_IMPL(double)
 }  // namespace onnxruntime