--- a/onnxruntime/core/providers/cpu/tensor/copy.h
+++ b/onnxruntime/core/providers/cpu/tensor/copy.h
@@ -1,14 +1,24 @@
 #include "core/providers/common.h"
 #include "core/platform/threadpool.h"
+#include <vector>
 
 namespace onnxruntime {
 
-template <typename T>
-void StridedCopy(onnxruntime::concurrency::ThreadPool* thread_pool,
-                 T* dst,
-                 std::vector<int64_t> dst_shape,
-                 std::vector<int64_t> dst_strides,
-                 T* src,
-                 std::vector<int64_t> src_strides);
+Status DispatchStridedCopy(concurrency::ThreadPool* thread_pool,
+                           Tensor& dst,
+                           std::ptrdiff_t dst_offset,
+                           const std::vector<int64_t> dst_strides,
+                           const TensorShape& copy_shape,
+                           const Tensor& src,
+                           const std::vector<int64_t> src_strides);
 
+template <typename T>
+void StridedCopy(concurrency::ThreadPool* thread_pool,
+                 T* dst,
+                 const TensorShape& dst_shape,
+                 const std::vector<int64_t>& dst_strides,
+                 const T* src,
+                 const std::vector<int64_t>& src_strides);
+
+std::vector<int64_t> StridesForTensor(const Tensor& tensor);
 }  // namespace onnxruntime