**Description**: 
Adds a `StridedCopy` function that implements a copy from strided tensor to another.

**Motivation and Context**
- This parallelizes the Concat operator, and can also be used in future to parallelize many other data movement operators (e.g. Transpose, Split, etc.).
- This operation is also required for the proposed data layout extensions to ORT. 

**Performance Characteristics**
All benchmarks measure the concatenation of two tensors with feature size 64 and variable batch size.

The PR introduces some overhead at smaller batch sizes: for axis 0, the new PR is at ~most half as slow as the current master with batch sizes less than 2048~ (not anymore, see below). For batch sizes above that, the performance is roughly similar (0.9x-1.05x). 

For multithreaded concatenations, speedup vs master scales slightly less than linearly with the number of cores, as would be expected. Of the (small) tested grid of parameters, the highest measured speedup is ~8x with 12 threads. 
At higher thread counts and high batch sizes, we also measure noticeable (1.5x-6x) speedups vs the parallelized PyTorch implementation.

(Please ping me for the detailed numbers)