@orausch 2021-07-09
Note: since we did decide to implement a more generalized tensor iterator, this op will likely be better implemented using that abstraction in the future.