@orausch 2021-07-21
> I don't have a clear picture on if that is likely to occur in practice. If it occurs in practice, or we need to be cautious, then can we keep the basic original code, and call into it below a threshold?

I've now added a "fast path" for the 2 dimensional case that is similar to what ORT is was doing before. ORT was previously able to take this path in all concat cases because tensors were always contiguous, but strided copy needs to support more cases than that. 

This path is taken when tensor is <2D. To maximize the cases where it is hit, I've also added contiguous dimension coalescing that reduces the dimensions of the tensors as much as possible.
<details>
<summary> Here is the single threaded profiling after this change (all times in us, median of 500 runs) </summary>

batch_size | feature_size | branch_median | master_median | Speedup Branch vs Master
-- | -- | -- | -- | --
32 | 64 | 3 | 2 | 0.666667
64 | 64 | 3 | 3 | 1
128 | 64 | 4 | 3 | 0.75
256 | 64 | 5 | 5 | 1
512 | 64 | 8 | 8 | 1
1024 | 64 | 22 | 21 | 0.954545
2048 | 64 | 73 | 72 | 0.986301
4096 | 64 | 158.5 | 158 | 0.996845
8192 | 64 | 294 | 286 | 0.972789
16384 | 64 | 614 | 588 | 0.957655
32768 | 64 | 1981.5 | 2061.5 | 1.040373
65536 | 64 | 6308.5 | 6114 | 0.969169

</details>